+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 
 - Removed the previously deprecated `ConvertibleColor` typealias.
 - Removed the `CssColors` object. Use `Color.fromCss` instead. 
+- Removed `Ansi16` companion object color constants.
 
 ## [2.1.0] - 2021-05-05
 ### Added